% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTPlotting_Chromosome.R
\name{GTPlotting_Chromosome}
\alias{GTPlotting_Chromosome}
\title{Generate the genotype plots for each chromosome in seprate files.}
\usage{
GTPlotting_Chromosome(vcfTablefile, chrome_length, ctrl_names)
}
\arguments{
\item{vcfTablefile}{Dataframe which contains the important features extracted from a vcf file using VcfToTable function.}

\item{chrome_length}{Dataframe which contains the size of each chromosome which is the output of the VcfToTable function.}

\item{ctrl_names}{String which gives the name of the Controlled sample should be the same as in the vcf file.}
}
\value{
genotype plots for each chromosome
}
\description{
Generate the genotype plots for each chromosome in seprate files.
}
