% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTPlotting_Chromosome_Combined.R
\name{GTPlotting_Chromosome_Combined}
\alias{GTPlotting_Chromosome_Combined}
\title{Generate the genotype plots for all the chromosomes in single plot.}
\usage{
GTPlotting_Chromosome_Combined(vcfTablefile, chrome_length)
}
\arguments{
\item{vcfTablefile}{Dataframe which contains the important features extracted from a vcf file using VcfToTable function.}

\item{chrome_length}{Dataframe which contains the size of each chromosome which is the output of the VcfToTable function.}
}
\value{
genotype plot for all the chromosome
}
\description{
Generate the genotype plots for all the chromosomes in single plot.
}
