% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VcfToTable.R
\name{VcfToTable}
\alias{VcfToTable}
\title{Converts the VCF file into a dataframe by extracting the required features for Genotype plotting.}
\usage{
VcfToTable(vcffile)
}
\arguments{
\item{vcffile}{String specifying name of VCF file in .vcf or .vcf.gz format.}
}
\value{
Object which consists of two Data Frames the first one vcfdata contains the VCF DataFrame which consists of the features required
    for genotype plotting and the second dataframe consists of the length of the different Chromosomes in the VCF file.
}
\description{
Converts the VCF file into a dataframe by extracting the required features for Genotype plotting.
}
